net: lib: nrf_cloud_pgps: Add full prediction storage injection test
Add nrf_cloud_injection_test() to quickly inject all predictions
to modem to confirm they are valid.

Call this from MMSS when button 1 is pressed (after any residual
delay from CONFIG_SENSOR_SAMPLE_INTERVAL_SECONDS) as a proof of
concept, with the damage_first parameter set false, to use the data
as provided by the cloud.

Call this from MMSS when button 2 is pressed (after any residual
delay from CONFIG_SENSOR_SAMPLE_INTERVAL_SECONDS) as a proof of
concept, with the damage_first parameter set true, to simulate
the corrupted data reported in
https://projecttools.nordicsemi.no/jira/browse/NCSDK-19089
in order to confirm errors will be reported in the logs.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>